fix: 금일 판매 금액이 갱신되도록 수정
상품 판매 시 바로 api 호출하도록 변경함.
